requirements.txt
rich
openai
google-generativeai